[ExeFramework] Rename packages in new debugger plugins
