+++ framework/execution_framework/plugins/org.executionframework.debugger.ui/src/org/executionframework/debugger/ui/MutableFieldNameProvider.xtend
