+++ framework/execution_framework/plugins/org.executionframework.debugger.ui/src/org/executionframework/debugger/ui/breakpoints/GemocToggleBreakpointAction.java
@@ -8,13 +8,13 @@
  * Contributors:
  *     Inria - initial API and implementation
  *******************************************************************************/
-package org.gemoc.executionframework.engine.ui.debug.sirius.action;
+package org.executionframework.debugger.ui.breakpoints;
 
 import org.eclipse.core.runtime.CoreException;
 import org.eclipse.emf.ecore.EObject;
 import org.eclipse.emf.ecore.resource.Resource;
 import org.eclipse.emf.ecore.resource.ResourceSet;
-import org.gemoc.executionframework.engine.ui.debug.breakpoint.GemocBreakpoint;
+import org.executionframework.debugger.ui.breakpoint.GemocBreakpoint;
 import org.gemoc.executionframework.engine.ui.launcher.AbstractGemocLauncher;
 
 import fr.inria.diverse.melange.resource.MelangeResource;
