+++ framework/execution_framework/plugins/org.executionframework.debugger/META-INF/MANIFEST.MF
@@ -4,7 +4,7 @@ Bundle-Name: Debugger
 Bundle-SymbolicName: org.executionframework.debugger
 Bundle-Version: 1.0.0.qualifier
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
-Export-Package: org.gemoc.executionframework.engine.ui.debug;uses:="org.eclipse.emf.ecore",
+Export-Package: org.executionframework.debugger;uses:="org.eclipse.emf.ecore",
  org.gemoc.executionframework.engine.ui.debug.breakpoint;
   uses:="org.eclipse.emf.ecore,
    org.eclipse.ui,
