+++ framework/execution_framework/plugins/org.executionframework.debugger/src/org/executionframework/debugger/IntrospectiveMutableFieldExtractor.xtend
